Added error messaging and more auth

+++ backend/routes/posts.js
@@ -35,9 +35,9 @@ multer({storage: storage}).single('image'),
     title: req.body.title,
     content: req.body.content,
     imagePath: url + '/images/' + req.file.filename,
-    postTime: req.body.postTime
+    postTime: req.body.postTime,
+    creator: req.userData.userId
   });
-
   post.save()
     .then(createdPost => {
     res.status(201).json({
@@ -47,10 +47,15 @@ multer({storage: storage}).single('image'),
         title: createdPost.title,
         content: createdPost.content,
         imagePath: createdPost.imagePath,
-        postTime: createdPost.postTime
+        postTime: createdPost.postTime,
       }
     });
-  });
+  })
+  .catch(error => {
+    res.status(500).json({
+      message: 'Creating a post failed!'
+    })
+  })
 });
 
 router.put('/:id',
@@ -67,12 +72,22 @@ multer({storage: storage}).single('image'),
     title: req.body.title,
     content: req.body.content,
     imagePath: imagePath,
-    postTime: req.body.postTime
+    postTime: req.body.postTime,
+    creator: req.userData.userId
   });
 
-  Post.updateOne({_id: req.params.id}, post).then(result => {
-    res.status(200).json({message: 'Update successful!'})
-  });
+  Post.updateOne({_id: req.params.id, creator: req.userData.userId}, post).then(result => {
+    if(result.nModified > 0){
+      res.status(200).json({message: 'Update successful!'});
+    } else {
+      res.status(401).json({message: 'Not Authorized!'});
+    }
+  })
+  .catch(error => {
+    res.status(500).json({
+      message: 'Couldn\'t update post!'
+    })
+  })
 });
 
 router.get('', (req, res, next) => {
@@ -94,6 +109,11 @@ router.get('', (req, res, next) => {
         maxPosts: count
       });
     })
+    .catch(error => {
+      res.status(500).json({
+        message: 'Fetching posts failed!'
+      })
+    })
 });
 
 router.get('/:id', (req, res, next) => {
@@ -104,13 +124,26 @@ router.get('/:id', (req, res, next) => {
       res.status(404).json({message: 'Post not found'});
     }
   })
+  .catch(error => {
+    res.status(500).json({
+      message: 'Fetching post failed!'
+    })
+  })
 });
 
 router.delete('/:id', checkAuth, (req, res, next) => {
-  Post.deleteOne({_id: req.params.id}).then(result => {
-    console.log(result);
-    res.status(200).json({message: 'Post Deleted'});
-  });
+  Post.deleteOne({_id: req.params.id, creator: req.userData.userId}).then(result => {
+    if(result.n > 0){
+      res.status(200).json({message: 'Update successful!'});
+    } else {
+      res.status(401).json({message: 'Not Authorized!'});
+    }
+  })
+  .catch(error => {
+    res.status(500).json({
+      message: 'Fetching posts failed!'
+    })
+  })
 })
 
 module.exports = router;
